+++ lab5/matrix.h
@@ -91,7 +91,7 @@ class matrix
 
 		static matrix translation(int dx, int dy);
 		static matrix rotation(int theta);
-		static matrix scale(int x, int y);
+		static matrix scale(double x, double y);
 
 		/**
 		 * @brief Creates a coordinate matrix/vector
