+++ lab5/viewcontext.h
@@ -29,7 +29,7 @@ class ViewContext
         std::unique_ptr<matrix> get_model_coords(const std::unique_ptr<matrix> device_coords) const;
 
         void translate(int dx, int dy);
-        void scale(int scale_x, int scale_y);
+        void scale(double scale_x, double scale_y);
 };
 
 #endif
